+++ src/main/java/com/b1n_ry/yigd/compat/InventorioCompat.java
@@ -80,7 +80,7 @@ public class InventorioCompat implements InvModCompat<DefaultedList<Pair<ItemSta
         }
 
         @Override
-        public DefaultedList<ItemStack> merge(CompatComponent<?> mergingComponent) {
+        public DefaultedList<ItemStack> merge(CompatComponent<?> mergingComponent, ServerPlayerEntity merger) {
             DefaultedList<ItemStack> extraItems = DefaultedList.of();
 
             @SuppressWarnings("unchecked")